refactor: effects from PNX 2.0

+++ src/main/java/cn/nukkit/nbt/tag/CompoundTag.java
@@ -267,7 +267,7 @@ public class CompoundTag extends Tag implements Cloneable {
 
     public int getShort(String name) {
         if (!tags.containsKey(name)) return 0;
-        return ((NumberTag) tags.get(name)).getData().intValue();
+        return ((NumberTag) tags.get(name)).getData().shortValue();
     }
 
     public int getInt(String name) {
